Refactor specific parameters folder layout
