+++ packages/ui/src/proposals/modals/AddNewProposal/components/SpecificParameters/SpecificParametersStep.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { State, Typestate } from 'xstate'
 
-import { CreateWorkingGroupLeadOpening } from '@/proposals/modals/AddNewProposal/components/SpecificParameters/CreateWorkingGroupLeadOpening'
 import { FundingRequest } from '@/proposals/modals/AddNewProposal/components/SpecificParameters/FundingRequest'
+import { CreateWorkingGroupLeadOpening } from '@/proposals/modals/AddNewProposal/components/SpecificParameters/WorkingGroupLeadOpening/CreateWorkingGroupLeadOpening'
 import {
   AddNewProposalContext,
   AddNewProposalEvent,
